Add utilization and missing_hosts database integration
Enable a ResourceQuota to track persistently which host resources were
missing during the last calculation of its utilization.

Extend the database with fields for utilization_<resource> and
missing_hosts. Add a one-to-many relation between Hosts and
ResourceQuotas, as a ResourceQuota can have several Hosts whose
resources could not be determined.

* Add ResourceQuotaMissingHost model
* Exclude ResourceQuota from Metrics/ClassLength
* Add api/v2/id/missing_hosts endpoint
* Add controller tests for missings_hosts and utilization
* Add model tests for missing_hosts and utilization
* Move generic stub functions to test_plugin_helper

Co-authored-by: Nadja Heitmann <nadjah@atix.de>