Implement optimistic resource quota utilization
Instead of re-computing the Resource Quota utilization by iterating all
Resource Quota hosts and determining their resources, an optimistic tracking
of the utilized resources is implemented.

When a host is created/added to a Resource Quota, its resources are
determined once and added to the Resource Quota utilization. When a
host is destroyed/unassigned from a Resource Quota, its utilization is
freed from the host's resource capacities.