Add RecurringLogic to update HostResources
Add task RefreshResourceQuotaUtilization which calls determine_utilization
on every ResourceQuota. The command iterates all hosts of a ResourceQuota
and sets the host.host_resources which is used for the ResourceQuota.utilization.

This guarantees that Foreman does not miss changes to HostResources,
when they are not performed via the Foreman UI. Moreover, it increases
performance since Foreman Resource Quota does not re-compute a
ResourceQuota's utilization on host deployment and, instead, takes
the last-computed sum of HostResources.

However, this comes with the downside that HostResources might be missed
when edited out-of Foreman UI and new hosts are deployed before the
task is executed.