Replace float error with rounding check
The React component 'UnitInputField' prints an error
if a user wants to type a value which is a float in the
corresponding base unit, for example 5.5 MiB for the memory limit.
Instead of showing an error, we introduce a warning which says
that the value is rounded to the next lower value (5.0 MiB).