Add 'Unassigned' quota as default quota
The initial implementation works without the need to set a quota -
there is a setting 'resource_quota_optional_assignment' which can be
set to true such that setting the quota becomes optional when creating
a host. This is to leverage possible negative user experience when a
quota is exceeded, especially at an early stage of plugin development.

In the initial design, for host registration the check implementation
does not work for mandatory assignment of a quota. In general, it is
possible to have no resource quota assigned, hence the quota can be set
to 'nil'. This works well for the optional assignment in general and
also host deployment via orcharhino with mandatory assignment. However,
host registration with mandatory assignment fails. This is because
there is a check when saving a host that fires when a resource quota
should be set but is not. There are two problems:

1) At the moment, there is no possibility to set the resource quota
during host registration, so this check always fails.

2) Even when the quota is set, there is a host creation process in
Katello in such an early stage that no quota is set at that point and
the check still fails even if a quota is provided because the quota
check fires on host.save! .

This PR alleviates the problem during host registration by
introducing a new quota 'Unassigned' which is set at host creation
when no quota is assigned. However, this requires a significant
redesign of the underlying plugin structure.